Split up functions to be more modular
